PIV-68: Fixed incorrect type

+++ ykcs11/ykcs11.c
@@ -830,7 +830,7 @@ CK_DEFINE_FUNCTION(CK_RV, C_OpenSession)(
       piv_obj_id_t atst_id = find_atst_object(sub_id);
       CK_ULONG slot = piv_2_ykpiv(pvtk_id);
       CK_BYTE data[YKPIV_OBJ_MAX_SIZE];  // Max cert value for ykpiv
-      unsigned long len;
+      size_t len;
       if(pvtk_id != PIV_INVALID_OBJ) {
         len = sizeof(data);
         if((rc = ykpiv_get_metadata(session->slot->piv_state, slot, data, &len)) == YKPIV_OK) {
